post-service/requirements.txt
flask
flask_sqlalchemy
werkzeug
PyJWT
requests